# lib/api-zod/src/generated/types/purchaseItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for a simple private inventory and sales app
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseItem {
  productId: number;
  productName: string;
  /** Código del producto al momento de la compra */
  productCode?: string | null;
  /** Contenido / presentación del producto al momento de la compra */
  content?: string;
  quantity: number;
  unitCost: number;
  subtotal: number;
}